+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to the "todo-tools" extension will be documented in this file.
 
 ## [0.8.9]
+### add
+- recur(weekday)
+
 
 ## [0.8.8] - 2025-06-18
 ### add